Rearranged file structure: created lib folder
