+++ lib/http_codes.rb
@@ -1,0 +1,2 @@
+require 'http_codes/httpcodes.rb'
+require 'http_codes/version.rb'
