+++ lib/http_codes/version.rb
@@ -1,0 +1,6 @@
+# Versioning
+
+module HttpCodesFetcher
+  VERSION = '0.1.0'
+  DATE = '2015-11-05'
+end
